% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare and extract seasonal summaries of tracked individuals during breeding season to predict nesting behaviour.}
\usage{
data_prep(
  trackingdata,
  indseasondata,
  latboundary = 45,
  longboundary = 4,
  minlocs = 800,
  nestradius = 50,
  homeradius = 2000,
  startseason = 70,
  endseason = 175,
  broodstart = yday(ymd("2023-05-01")),
  broodend = yday(ymd("2023-06-01")),
  settleEnd = 97,
  Incu1End = 113,
  Incu2End = 129,
  Chick1End = 152,
  age = 10
)
}
\arguments{
\item{trackingdata}{data.frame with tracking data set with information on individual season and id as unique identifier ('year_id').
Must contain a timestamp, and coordinates in both geographic ('lat_wgs', 'long_wgs') and projected coordinate systems ('lat_eea', 'long_eea').}

\item{indseasondata}{data.frame with nesting information per individual and season, which links to \code{trackingdata} by the unique identifier ('year_id').
Must contain columns for the age ('age_cy' as integer in calender years) and sex ('sex': either 'm' or 'f') of individuals in that season.
Will optionally contain three columns of binary training data, namely 'HR' (whether a home range existed or not), 'nest' (whether a nesting attempt took place) and 'fledged' (whether the nest was successful)
If no nesting information is available, \code{indseasondata} can be created from \code{trackingdata} by grouping and summarising \code{trackingdata} to yield one line per individual and season.}

\item{latboundary}{numeric. Value for the southern latitudinal boundary, i.e. locations will only be retained if \code{trackingdata$lat_wgs} is >\code{latboundary}}

\item{longboundary}{numeric. Value for the western longitudinal boundary, i.e. locations will only be retained if \code{trackingdata$long_wgs}  is >\code{longboundary}}

\item{minlocs}{integer. Minimum number of GPS locations that an individual must have recorded between \code{startseason} and \code{endseason} to be considered for analysis. The remaining individuals will be removed.}

\item{nestradius}{numeric. Radius around potential nest site in metres.
This radius will determine in which radius the residency times will be calculated, i.e. how long an animal remained within that radius or how long it spent outside this radius.
Please consider device accuracy and general movement behaviour of the species when specifying the radius. Default is 50 m.}

\item{homeradius}{numeric. Radius of home ranging movements around potential nest site in metres.
This radius will determine the core home range and will be used to quantify the amount of time individuals spent outside their core home range to inform the distinction of nest failure.
Please consider general movement behaviour of the species when specifying the radius. Default is 2000 m.}

\item{startseason}{integer. Day of the year (obtained by \code{yday(SomeDate)}) that marks the start of the breeding season. Tracking data before this date will be removed.
This day should be the earliest recorded egg laying date for the species. Defaults to 70 (11 Mar).}

\item{endseason}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the breeding season. Tracking data after this date will be removed.
This day should be the earliest recorded fledging date for the species. Defaults to 175 (24 Jun).}

\item{broodstart}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the start of the small chick phase. Defaults to 121 (1 May).}

\item{broodend}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the small chick phase. Defaults to 152 (1 Jun).}

\item{settleEnd}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the settlement phase at the beginning of the breeding season.
This day will demarcate the time period when birds exhibit a different movement behaviour than once eggs have been laid. Defaults to 97 (7 Apr).}

\item{Incu1End}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the first half of the incubation phase during the breeding season.
This day will facilitate that different movement behaviours are quantified during incubation. Defaults to 113 (23 Apr).}

\item{Incu2End}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the incubation phase during the breeding season.
This day will demarcate the time period when birds exhibit a different movement behaviour once chicks hatch. Defaults to 129 (9 May).}

\item{Chick1End}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the first half of the chick rearing phase during the breeding season.
This day will facilitate that different movement behaviours are quantified during chick rearing. Defaults to 152 (1 Jun).}

\item{age}{integer. For birds that have no age information, or for datasets where no 'age_cy' (age in calender years) exists, a value will be filled in to facilitate the prediction of nest success.
Defaults to 10, but should be informed by a sensible value for the species in question, so that \code{age} reflects the age in calendar years of birds for which no age information exists.}
}
\value{
Returns a list with 4 data.frames. \code{summary} contains one row per individual season and 45 variables, which each characterise movement metrics for a different period of the breeding season. Output data.frame includes age and sex information.
\code{pot_nests} contains one row per individual season and two variables (x and y), representing the longitude and latitude of potential nest sites (in projected coordinate system).
\code{nightlocs} contains all nocturnal locations for all individuals, including the number of revisits to within \code{nestradius} and the residence time within \code{nestradius} at that location.
\code{daylocs} contains all diurnal locations for all individuals, including the number of revisits to within \code{nestradius} and the residence time within \code{nestradius} at that location.
}
\description{
\code{data_prep} extracts temporally varying summary information from tracking data of individuals
to help diagnose whether a breeding attempt was initiated, and whether it was successful or not. The following metrics are calculated: 95\% MCP home range area,
maximum time away from nest radius, median daily time spent at nest, median daily travel distance, and median daily maximum distance from nest.
The function first identifies the most plausible location of a potential nest by calculating time spent and revisits to all locations within a buffer of
\code{nestradius}, and then uses this location of a nest to calculate distances and times spent within or outwith that radius.
}
