% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_metric_extraction.R
\name{move_metric_extraction}
\alias{move_metric_extraction}
\title{Extract seasonal movement metrics of tracked individuals during breeding season}
\usage{
move_metric_extraction(
  trackingdata,
  nest_locs,
  inddata,
  nestradius = 50,
  startseason = 70,
  endseason = 175,
  uncertainty = 0.45
)
}
\arguments{
\item{trackingdata}{tibble with tracking data set with information on individual season and id.
Created by \code{\link{data_prep}} and returned as \code{movementtrack}. Must contain id, timestamp, and coordinates in projected coordinate system.}

\item{nest_locs}{data.frame with potential nest coordinates for each individual season and id.
Created by \code{\link{data_prep}} and returned as \code{pot_nests}. Must contain id, and nest coordinates in projected coordinate system.}

\item{inddata}{data.frame with unique identifier for individual seasons and predicted probability of nest success provided function \code{\link{predict_success}}.
Must contain columns for the age and sex of individuals in that season, and a column of \code{succ_prob} as provided by the function \code{\link{predict_success}}. Format of id must match that of \code{trackingdata}}

\item{nestradius}{numeric. Radius around potential nest site in metres.
This radius will determine in which radius times will be calculated, i.e. how long an animal remained within that radius or how long it spent outside this radius.
Please consider device accuracy and general movement behaviour of the species when specifying the radius.}

\item{startseason}{integer. Day of the year (obtained by \code{yday(SomeDate)}) that marks the start of the breeding season. Tracking data before this date will be removed.
This day should be the earliest recorded egg laying date for the species. Defaults to 70 (11 Mar).}

\item{endseason}{integer. Day of the year (obtained by \code{yday(SomeOtherDate)}) that marks the end of the breeding season. Tracking data after this date will be removed.
This day should be the earliest recorded fledging date for the species. Defaults to 175 (24 Jun).}

\item{uncertainty}{numeric value between 0 and 0.5. Individuals for which the nest success classification resulted in a probability <\code{uncertainty}
are retained for data extraction. A value close to 0 will result in many individuals falling below the uncertainty threshold, resulting in longer runtime of the function as more individuals will be processed.
A value close to 0.5 will result in very few individuals that will be processed and for which movement metrics will be returned.}
}
\value{
Returns a data.frame with movement metrics for each individual season and 5-day window, including age and sex information.
}
\description{
\code{move_metric_extraction} extracts temporally varying summary information from tracking data of individuals
to help diagnose whether a breeding attempt was successful or not. The following metrics are calculate: 95\% MCP home range area,
maximum time away from nest radius, median daily time spent at nest, median daily travel distance, and median daily maximum distance from nest.
The function first identifies the most plausible location of a potential nest by calculating time spent and revisits to all locations within a buffer of
\code{nestradius}, and then uses this location of a nest to calculate distances and times spent within or outwith that radius.
}
