% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_visualisation.R
\name{movement_visualisation}
\alias{movement_visualisation}
\title{Visualise seasonal movement of tracked individuals during breeding season in an interactive shiny app}
\usage{
movement_visualisation(
  trackingdata,
  nest_locs,
  inddata,
  move_metrics,
  uncertainty_success = 0.25,
  uncertainty_nest = 0.25,
  output_path = "output/05_full_run_CH/nest_success_output.csv"
)
}
\arguments{
\item{trackingdata}{filename for data.frame with tracking data derived from \code{\link{data_prep}} as \code{nest_data_input$movementtrack}. This data.frame is used to visualise the GPS locations and the movement trajectory on the map.}

\item{inddata}{filename for data.frame with predictions for breeding success derived from \code{\link{predict_success}} as \code{pred_succ}. This data.frame is used to visualise brood relevant metrics as table.}

\item{move_metrics}{filename for data.frame with seasonal movement metrics during breeding season derived from \code{\link{move_metric_extraction}} as \code{milvus_5d_move_metrics}. This data.frame is used to plot the move metrics.}

\item{output_path}{path for the data.frame with the stored decisions about the examined individuals, whether they have had a nest and a successful brood.}

\item{uncertainty}{numeric value between 0 and 0.5. Individuals for which the nest success classification resulted in a probability <\code{uncertainty}. This value should match the value of \code{uncertainty} used in \code{\link{move_metric_extraction}} to ensure that data is available for all visualised individuals.}
}
\value{
Returns an interactive shiny app that shows a map, a table and a plot for every analysed individual.
}
\description{
\code{movement_visualisation} visualises tracking data of individuals on an interactive map
to help diagnose whether a breeding attempt has been made and whether it was successful or not.
The decisions whether a breeding attempt has taken place and whether it was successful are
storedin an external data frame and visualised as a table in a panel of the shiny app. To
support the decision, in a separate panel \code{plot_move_metrics} is integrated which produces
a series of plots of movement metrics summarised over a moving time window for the analysed individual.
}
\details{
Input data for this function must be derived from \code{\link{data_prep}} (\code{nest_data_input$movementtrack} and \code{nest_data_input$pot_nests}),
\code{\link{predict_success}} (\code{pred_succ}) and \code{\link{move_metric_extraction}} (\code{milvus_5d_move_metrics})
}
