% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestTool-package.R
\docType{package}
\name{NestTool-package}
\alias{NestTool}
\alias{NestTool-package}
\title{NestTool: Detect nesting and outcome from tracking data}
\description{
The NestTool package is an informal collection of R functions created by the Swiss Ornithological Institute to extract demographic information from the tracking data of birds. The functions prepare movement metrics, and predict whether a nesting attempt occurred and whether the attempt was successful.
}
\author{
\strong{Maintainer}: Steffen Oppel \email{steffen.oppel@vogelwarte.ch} (\href{https://orcid.org/0000-0002-8220-3789}{ORCID})

}
\keyword{internal}
