% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_move_metrics.R
\name{plot_move_metrics}
\alias{plot_move_metrics}
\title{Plot seasonal movement metrics of tracked individuals during breeding season}
\usage{
plot_move_metrics(movemetrics, individual)
}
\arguments{
\item{movemetrics}{data.frame derived from \code{\link{move_metric_extraction}}.
This table will have multiple columns for different movement metrics, and a column for the individual ("id"),the time period ("week")
and age and sex of the respective individual in that season.}

\item{individual}{character. Unique identifier for individual seasons.
Must be contained in column "id" of \code{movemetrics}}
}
\value{
Returns a ggplotly plot and writes a pdf version of the plot into a directory.
}
\description{
\code{plot_move_metrics} produces a series of plots of movement metrics summarised over a moving time window
to help diagnose whether a breeding attempt was successful or not.
}
\details{
Input data for this function must be derived from \code{\link{move_metric_extraction}}
}
