% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nesting.R
\name{predict_nesting}
\alias{predict_nesting}
\title{Predict whether nesting occurred based on movement summary metrics.}
\usage{
predict_nesting(
  model = readRDS(here::here("data/nest_model.rds"))$model,
  trackingsummary
)
}
\arguments{
\item{model}{RandomForest model object created with \code{\link{train_nest_detection}}. If none is provided, a model trained for red kites in Switzerland is used.}

\item{trackingsummary}{data.frame created by \code{\link{data_prep}} with information on individual identity, age, sex, and seasonal movement metrics.
If this file results from the output of \code{\link{predict_nesting}}, then no further prediction will be performed as the prediction is already contained in the file.}
}
\value{
Returns a data.frame with predicted probabilities of a nest, including all input data for further use in \code{\link{predict_success}}.
}
\description{
\code{predict_nesting} uses temporally varying summary information from tracking data of individuals
to predict whether a breeding attempt occurred or not. The input data will have been prepared by \code{\link{data_prep}}, and a model may have been created by \code{\link{train_nest_detection}} if suitable training data exist from the
studied population. If no training data exist, and no model has been created, the function will draw on a model created by the Swiss Ornithological Institute for Red Kites tracked in Switzerland.
The Random Forest model either created by \code{\link{train_nest_detection}} or contained in the package will be used to predict the probability that a given individual in the season initiated a nesting attempt or not.
}
