% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_success.R
\name{predict_success}
\alias{predict_success}
\title{Predict whether nesting was successful based on movement summary metrics.}
\usage{
predict_success(
  model = readRDS(here("R/succ_model.rds"))$model,
  nestingsummary,
  nest_cutoff = readRDS(here("R/succ_model.rds"))$nest_cutoff
)
}
\arguments{
\item{model}{RandomForest model object created with \code{\link{train_nest_success}}. If none is provided, a model trained for red kites in Switzerland is used.}

\item{nestingsummary}{data.frame created by \code{\link{predict_nesting}} with information on individual identity, age, sex, seasonal movement metrics, and probability of nesting.
If this file results from the output of \code{\link{train_nest_success}}, then no further prediction will be performed as the prediction is already contained in the file.}

\item{nest_cutoff}{numeric between 0 and 1. Created by \code{\link{train_nest_success}} to indicate the minimum probability of a nesting attempt occurring in the training data for \code{model}
All nesting attempts that have a 'nest_prob' below this threshold will automatically be classified as unsuccessful.}
}
\value{
Returns a data.frame with predicted probabilities of nest success.
}
\description{
\code{predict_success} uses temporally varying summary information from tracking data of individuals
to predict whether a breeding attempt was successful or not. The input data will have been prepared by \code{\link{predict_nesting}}, and a model may have been created by \code{\link{train_nest_success}} if suitable training data exist from the
studied population. If no training data exist, and no model has been created, the function will draw on a model created by the Swiss Ornithological Institute for Red Kites tracked in Switzerland.
The Random Forest model either created by \code{\link{train_nest_success}} or contained in the package will be used to predict the probability that a given individual in the season successfully bred or not.
}
