% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_nest_detection.R
\name{train_nest_detection}
\alias{train_nest_detection}
\title{Train a model that predicts whether nesting occurred based on movement summary metrics.}
\usage{
train_nest_detection(trackingsummary, plot = TRUE)
}
\arguments{
\item{trackingsummary}{data.frame created by \code{\link{data_prep}} with information on individual identity, age, sex, and seasonal movement metrics.
Must contain a column 'nest' with two possible values, 'no nest' or 'nest' (or 0 or 1) to be able to train the model (at least 5 of each are required). If no such data are available, use \code{\link{predict_nesting}} instead.}

\item{plot}{logical. If TRUE, a variable importance plot is produced.}
}
\value{
Returns a list with four elements: \code{model} is the random forest model to predict nesting;
\code{summary} is the output data.frame with predicted probabilities of a nest, including all input data for further use in \code{\link{predict_success}};
\code{eval_train} is a confusion matrix and accuracy assessment of the predictive accuracy of the model on training data.
\code{eval_test} is a confusion matrix and accuracy assessment of the predictive accuracy of the model on internally cross-validated test data.
}
\description{
\code{train_nest_detection} uses temporally varying summary information from tracking data of individuals
to help diagnose whether a breeding attempt occurred or not. The input data will have been prepared by \code{\link{data_prep}}, and are passed into a RandomForest model that uses the provided information on whether an individual had a nest or not in a given season.
The model is evaluated using internal cross-validation and a plot of variable importance and accuracy tables are returned.
}
