% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_visualisation.R
\name{movement_visualisation}
\alias{movement_visualisation}
\title{Visualise seasonal movement of tracked individuals during breeding season in an interactive shiny app}
\usage{
movement_visualisation(
  trackingdata,
  crs_epsg = 3035,
  nest_locs,
  inddata,
  move_metrics,
  uncertainty = 0.25,
  output_path = "NestTool_nest_success_output.csv"
)
}
\arguments{
\item{trackingdata}{data.frame with tracking data derived from \code{\link{data_prep}} as \code{nest_data_input$movementtrack}. This data.frame is used to visualise the GPS locations and the movement trajectory on the map.}

\item{crs_epsg}{numeric. EPSG code for the Coordinate Reference System of the projected coordinates used in \code{\link{data_prep}}. \link[=https://epsg.io/]{Find EPSG code here}. Default is 3035 (Lamberth Azimuthal Equal Area for Europe)}

\item{nest_locs}{data.frame with coordinates of potential nest locations for every individual derived from \code{\link{data_prep}} as \code{nest_data_input$pot_nests}. This data.frame is used to visualise the potential nest location on the map.}

\item{inddata}{data.frame with predictions for breeding success derived from \code{\link{predict_success}} as \code{pred_succ}. This data.frame is used to visualise brood relevant metrics as table.}

\item{move_metrics}{data.frame with seasonal movement metrics during breeding season derived from \code{\link{move_metric_extraction}}. This data.frame is used to plot the move metrics.}

\item{uncertainty}{numeric value between 0 and 0.5. Individuals for which the nest success classification resulted in a probability >\code{uncertainty} and <(1-\code{uncertainty}). This value should match the value of \code{uncertainty} used in \code{\link{move_metric_extraction}} to ensure that data is available for all visualised individuals.}

\item{output_path}{path for the data.frame with the stored decisions about the examined individuals, whether they have had a nest and a successful brood.}
}
\value{
Returns an interactive shiny app that shows a map, a table and a plot for every individual that falls above the absolute value of \code{uncertainty}. The user can then manually assign whether individuals initiated nesting and bred successfully. When the shiny app is closed, the manually annotated values will be saved in a csv file stored at \code{output_path}
}
\description{
\code{movement_visualisation} visualises tracking data of individuals on an interactive map
to help diagnose whether a breeding attempt has been made and whether it was successful or not.
The decisions whether a breeding attempt has taken place and whether it was successful are
storedin an external data frame and visualised as a table in a panel of the shiny app. To
support the decision, in a separate panel \code{plot_move_metrics} is integrated which produces
a series of plots of movement metrics summarised over a moving time window for the analysed individual.
}
\details{
Input data for this function must be derived from \code{\link{data_prep}} (\code{nest_data_input$movementtrack} and \code{nest_data_input$pot_nests}),
\code{\link{predict_success}} (\code{pred_succ}) and \code{\link{move_metric_extraction}}.
}
